shipping/requirements.txt
blinker==1.4
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
SQLAlchemy==1.4.27
Werkzeug==2.0.2
WTForms==3.0.1
